Correection du fait de la modification texte

+++ backend/controllers/message.js
@@ -54,22 +54,27 @@ exports.modifyMessage = async (request, response) => {
 		) {
 			return response.status(401).json({ message: 'Requete non autorisée' });
 		}
-		if (message.imageUrl !== undefined) {
-			const filename = message.imageUrl.split('/images/')[1];
-			fs.unlink(`images/${filename}`, () => {});
-		}
+
 		console.log('Request body');
 		console.log({ ...request.body, _id: request.params.id });
-		await Message.updateOne(
-			{ _id: request.params.id },
-			{
-				...request.body,
-				_id: request.params.id,
-				imageUrl: `${request.protocol}://${request.get('host')}/images/${
-					request.file.filename
-				}`,
+		const newMessage = {
+			...request.body,
+			_id: request.params.id,
+			imageUrl: message.imageUrl,
+		};
+		if (request.file !== undefined) {
+			if (message.imageUrl !== undefined) {
+				const filename = message.imageUrl.split('/images/')[1];
+				fs.unlink(`images/${filename}`, () => {});
 			}
-		);
+			const imageUrl = `${request.protocol}://${request.get('host')}/images/${
+				request.file.filename
+			}`;
+			newMessage.imageUrl = imageUrl;
+			console.log('Pas de modification');
+		}
+		console.log(newMessage);
+		await Message.updateOne({ _id: request.params.id }, newMessage);
 		response.status(200).json({ message: 'Objet modifié' });
 	} catch (e) {
 		console.log(e);
